ai-core/pyproject.toml
[project]
name = "ai-core"
version = "2.0.0"
description = "YOLOv8 Fire Detection with Region Localization"
authors = [
    {name = "Guilherme Soares",email = "soaresgui.dev@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # YOLOv8 Core Dependencies
    "ultralytics (>=8.3.191,<9.0.0)",
    "torch (>=2.0.0,<3.0.0)",
    "torchvision (>=0.15.0,<1.0.0)",
    "opencv-python (>=4.8.0,<5.0.0)",
    
    # Image Processing
    "pillow (>=11.3.0,<12.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    
    # Visualization & Plotting
    "matplotlib (>=3.10.6,<4.0.0)",
    "plotly (>=6.3.0,<7.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    
    # Dataset Management
    "roboflow (>=1.2.7,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    
    # Web & API Utils
    "requests (>=2.32.5,<3.0.0)",
    "streamlit (>=1.49.1,<2.0.0)",
    
    # Legacy Model Support (MobileNetV2)
    "tensorflow (>=2.20.0,<3.0.0)",
    "kagglehub (>=0.3.13,<0.4.0)",
]

[project.scripts]
fire-detect = "main:main"

[project.optional-dependencies]
gpu = [
    "torch[cuda]",
    "torchvision[cuda]",
]
dev = [
    "jupyter (>=1.0.0)",
    "pytest (>=7.0.0)",
    "black (>=23.0.0)",
    "flake8 (>=6.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
